Extract processing code
